refactor viewmodel
