+++ chat/src/main/java/cn/wildfire/chat/kit/channel/ChannelInfoActivity.java
@@ -23,7 +23,6 @@ import com.bumptech.glide.request.RequestOptions;
 import butterknife.Bind;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
-import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.common.OperateResult;
 import cn.wildfire.chat.kit.user.UserViewModel;
 import cn.wildfirechat.chat.R;
