+++ chat/src/main/java/cn/wildfire/chat/kit/contact/newfriend/InviteFriendActivity.java
@@ -11,7 +11,6 @@ import androidx.lifecycle.ViewModelProviders;
 import butterknife.Bind;
 import butterknife.OnClick;
 import cn.wildfire.chat.kit.WfcBaseActivity;
-import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.contact.ContactViewModel;
 import cn.wildfire.chat.kit.user.UserViewModel;
 import cn.wildfirechat.chat.R;
