+++ chat/src/main/java/cn/wildfire/chat/kit/conversation/SingleConversationInfoFragment.java
@@ -79,8 +79,8 @@ public class SingleConversationInfoFragment extends Fragment implements Conversa
     }
 
     private void init() {
-        conversationViewModel = ViewModelProviders.of(this, new ConversationViewModelFactory(conversationInfo.conversation)).get(ConversationViewModel.class);
-        userViewModel =ViewModelProviders.of(this).get(UserViewModel.class);
+        conversationViewModel = WfcUIKit.getAppScopeViewModel(ConversationViewModel.class);
+        userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
         String userId = conversationInfo.conversation.target;
         conversationMemberAdapter = new ConversationMemberAdapter(true, false);
         List<UserInfo> members = Collections.singletonList(userViewModel.getUserInfo(userId, false));
