+++ chat/src/main/java/cn/wildfire/chat/kit/conversation/forward/ForwardActivity.java
@@ -23,7 +23,6 @@ import java.util.List;
 import butterknife.Bind;
 import butterknife.OnTextChanged;
 import cn.wildfire.chat.kit.WfcBaseActivity;
-import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.common.OperateResult;
 import cn.wildfire.chat.kit.group.GroupViewModel;
 import cn.wildfire.chat.kit.search.OnResultItemClickListener;
