+++ chat/src/main/java/cn/wildfire/chat/kit/group/BasePickGroupMemberActivity.java
@@ -8,7 +8,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import cn.wildfire.chat.kit.WfcBaseActivity;
-import cn.wildfire.chat.kit.WfcUIKit;
 import cn.wildfire.chat.kit.contact.model.UIUserInfo;
 import cn.wildfire.chat.kit.contact.pick.PickUserViewModel;
 import cn.wildfire.chat.kit.user.UserViewModel;
